+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+- 2022-12-15
+  - Add an implementation for `greeting`
 
 ## [0.0.2]
 - 2022-12-15